+++ frontend/lib/screens/login_page.dart
@@ -1,40 +1,43 @@
 import 'package:flutter/material.dart';
 
-class LoginPage extends StatefulWidget {
-  const LoginPage({super.key});
-
-  @override
-  State<LoginPage> createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  final TextEditingController _idCtrl = TextEditingController();
-  final TextEditingController _pwCtrl = TextEditingController();
-  bool _keepLoggedIn = false;
+class LoginPage extends StatelessWidget {
+  const LoginPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('로그인')),
-      body: Padding(
-        padding: const EdgeInsets.all(16),
+      backgroundColor: Colors.white,
+      body: SafeArea(
         child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            TextField(controller: _idCtrl, decoration: const InputDecoration(labelText: '아이디')),
-            const SizedBox(height: 8),
-            TextField(controller: _pwCtrl, decoration: const InputDecoration(labelText: '비밀번호'), obscureText: true),
-            SwitchListTile(
-              title: const Text('로그인 상태 유지'),
-              value: _keepLoggedIn,
-              onChanged: (v) => setState(() => _keepLoggedIn = v),
+            // 상단 스페이서
+            const SizedBox(height: 1),
+
+            // 중앙 로고 텍스트
+            const Center(
+              child: Text(
+                'AutoToon',
+                style: TextStyle(
+                  fontSize: 28,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.black87,
+                ),
+              ),
             ),
-            ElevatedButton(
-              onPressed: () => Navigator.pushReplacementNamed(context, '/main'),
-              child: const Text('로그인'),
-            ),
-            TextButton(
-              onPressed: () => Navigator.pushNamed(context, '/signup'),
-              child: const Text('회원가입'),
+
+            // 하단 이미지 버튼
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 24),
+              child: GestureDetector(
+                onTap: () {
+                  Navigator.pushReplacementNamed(context, '/main');
+                },
+                child: Image.asset(
+                  'assets/images/kakao_login_medium_wide.png',
+                  fit: BoxFit.contain,
+                ),
+              ),
             ),
           ],
         ),
